requirements.txt
anthropic>=0.40.0
azure-identity>=1.16.0
azure-keyvault-secrets>=4.8.0
google-api-python-client>=2.131.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
